requirements.txt
Boto3==1.26.12
django==4.1.5
django-crispy-forms==1.14.0
django-storages==1.13.1
crispy-bootstrap5==0.7
dj-database-url>=0.5,<0.6
gunicorn==20.1.0
openai==0.26.4
Pillow==9.3.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0
